% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAM.R
\name{ADAM2.CalculateBayesianfactor}
\alias{ADAM2.CalculateBayesianfactor}
\title{Calculate the Log transformed Bayesian Factor given a bimodal distribution of gene ranks or slopes of the linear fit of ranks.}
\usage{
ADAM2.CalculateBayesianfactor(RankDistribution,
                display=TRUE,
                prefix='BayesianFactor')
}
\arguments{
\item{RankDistribution}{Quantative knockout screen dependency matrix where rows are genes and columns are samples. A real number in position \emph{[i,j]} represents the strength of dependency which indicates the amaount of loss of fitness in the \emph{j}-th sample in case of the inactivation of the \emph{i}-th gene. Higher strength of dependency indicates higher probability of beign a core fitness gene. These values are used for ranking the genes in terms of their dependecy strength.}

\item{display}{Boolean, default is TRUE. Should plots of the bimodal normal distribution fitting}

\item{prefix}{if the display is false the plots are generated in the working directory using the prefix.}
}
\value{
A data frame with the following columns:
\item{Gene}{Gene name}
\item{logBF}{Log of the Bayesian Factor}
}
\description{
This function calculates the log transformed bayesian factor between essential and non-essential genes, by fitting two normal distributins to the bimodal distribution of the ranks of genes in their \emph{N}-th percentile least dependent cell lines ordered wrt gene effect score/estimated slopes from linear fitting of the ordered ranks of genes in cell lines. The function fits a two normal distributions on the bimodal distribution and estimates the mean the standard deviation of these distributions. Based on the estimations the bayesian factor is calculated for each gene using the probabilities of genes coming from essential or non-essential distributions.
}
\details{
This function is using the mixdist R library to fit the given data into a distribution that can be represented as a mixture of two normal distributions. The mean and the standard deviations for each normal distribution is estimated
Each normal distribution corresponds to the essential and non-essential genes. For each gene a bayesian factor is calculated which can be defined as the Prob(gene|essential)/Prob(gene|non-essential). The log transformed bayesian factors are reported by this function.
}
\examples{
data(exampleSBFData)
results <- ADAM2.SlopeCF(depMat=exampleSBFData,display=TRUE)
bfresults <- ADAM2.CalculateBayesianfactor(RankDistribution=results$LeastDependent)
}
\author{
C. Pacini, E. Karakoc & F. Iorio
}
\keyword{functions}
