% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAM.R
\name{ADAM2.coreFitnessGenes}
\alias{ADAM2.coreFitnessGenes}
\title{Calculate the Core Fitness genes given the binary dependency matrix and the minimal number of cell line threshold.}
\usage{
ADAM2.coreFitnessGenes(depMat,
                              crossoverpoint)
}
\arguments{
\item{depMat}{Binary dependency matrix, rows are genes and columns are samples. 1 in position \emph{[i,j]} indicates that inactivation of the \emph{i}-th gene exerts a significant loss of fitness in the \emph{j}-th sample, 0 otherwise.}

\item{crossoverpoint}{minimum number of cell lines in which a gene needs to be fitness in order to be called core-fitness}
}
\value{
A vector that containing the Core Fitness Genes:
}
\description{
This function identifies the Core Fitness genes using the Adaptive Daisy Model (implemented ADAM) starting from a binary dependency matrix.
}
\details{
This function calculates the Core Fitness essential genes based on the calculated minimum number of cell lines that optimizes the True positive rates with log10 odds ratios. log10 odd ratios are calculated of observed vs. expected profiles of cumulative number of fitness genes in fixed number of cell lines. Expected values are the mean of those observed across randomised version of the observed binary matrix.
}
\examples{
data(exampleDepMat)
cfgenes <- ADAM2.coreFitnessGenes(depMat=exampleDepMat,crossoverpoint=3800)
}
\author{
C. Pacini, E. Karakoc & F. Iorio
}
\keyword{functions}
