% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAM.R
\name{ADAM2.coreFitnessGenesWrapper}
\alias{ADAM2.coreFitnessGenesWrapper}
\title{Calculate the Core Fitness genes, starting from the binary dependency matrix.}
\usage{
ADAM2.coreFitnessGenesWrapper(depMat,
                      display=TRUE,
                      main_suffix='fitness genes in at least 1 cell line',
                      xlab='n. dependent cell lines',
                      ntrials=1000)
}
\arguments{
\item{depMat}{Binary dependency matrix, rows are genes and columns are samples. 1 in position \emph{[i,j]} indicates that inactivation of the \emph{i}-th gene exerts a significant loss of fitness in the \emph{j}-th sample, 0 otherwise.}

\item{display}{Boolean, default is TRUE. Should bar plots of the dependency profiles be plotted}

\item{main_suffix}{If display=TRUE, title suffix to give to plot of number of genes depleted in a give number of cell lines, default is 'genes depleted in at least 1 cell line'}

\item{xlab}{label to give to x-axis of the plots, default is 'n. cell lines'}

\item{ntrials}{Integer, default =1000. How many times to randomly perturb dependency matrix to generate the null distributions.}
}
\value{
A vector that containing the Core Fitness Genes:
}
\description{
This function identifies the Core Fitness genes using the Adaptive Daisy Model (implemented ADAM) starting from a binary dependency matrix.
}
\details{
This function calculates the Core Fitness essential genes based on the calculated minimum number of cell lines that optimizes the True positive rates with log10 odds ratios. log10 odd ratios are calculated of observed vs. expected profiles of cumulative number of fitness genes in fixed number of cell lines. Expected values are the mean of those observed across randomised version of the observed binary matrix.
}
\examples{
data(exampleDepMat)
cfgenes <- ADAM2.coreFitnessGenesWrapper(depMat=exampleDepMat,ntrials=1000)
}
\author{
C. Pacini, E. Karakoc & F. Iorio
}
\keyword{functions}
