% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAM.R
\name{ADAM2.empiricalOdds}
\alias{ADAM2.empiricalOdds}
\title{Empirical odds of number of fitness genes per number of cell lines}
\usage{
ADAM2.empiricalOdds(observedCumSum,
                   simulatedCumSum)
}
\arguments{
\item{observedCumSum}{Observed profile of cumulative sum of numbers of fitness genes in fixed number of cell lines. This is generated by the \code{ADAM.panessprofile} function.}

\item{simulatedCumSum}{Random profiles of cumulative sum of fitness genes in fixed number of cell lines. This is generated by the function \code{ADAM.generateNullModel}.}
}
\value{
A named vector:
\item{odds}{log base 10 odd ratios of observed versus expected cumulative sums of number of fitness genes across fixed numbers of cell lines.}
}
\description{
This function calculates log10 odd ratios of observed vs. expected profiles of cumulative number of fitness genes in fixed number of cell lines. Expected values are the mean of those observed across randomised version of the observed binary matrix.
}
\examples{
data(exampleDepMat)
observed<-ADAM2.panessprofile(depMat=exampleDepMat)
null_m<-ADAM2.generateNullModel(depMat=exampleDepMat)
logOdds <- ADAM2.empiricalOdds(observedCumSum=observed$CUMsums,simulatedCumSum=null_m$nullCumSUM)
logOdds
}
\seealso{
ADAM2.panessprofile, ADAM2.generateNullModel
}
\author{
C. Pacini, E. Karakoc & F. Iorio
}
\keyword{functions}
