% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAM.R
\name{ADAM2.SlopeCF}
\alias{ADAM2.SlopeCF}
\title{Calculate the Core Fitness genes using linear modeling of the ranks of genes in all cell lines ordered wrt to their gene score effects}
\usage{
ADAM2.SlopeCF(depMat,
                display=TRUE,
                prefix='SlopeMethod')
}
\arguments{
\item{depMat}{Quantative knockout screen dependency matrix where rows are genes and columns are samples. A real number in position \emph{[i,j]} represents the strength of dependency which indicates the amount of loss of fitness in the \emph{j}-th sample in case of the inactivation of the \emph{i}-th gene. Higher strength of dependency indicates higher probability of beign a core fitness gene. These values are used for ranking the genes wrt to gene effect scores.}

\item{display}{Boolean, default is TRUE. Should bar plots of the dependency profiles be plotted}

\item{prefix}{if the display is false the plots are generated in the working directory using the prefix.}
}
\value{
A list of the following vectors:
\item{cfgenes}{Vector of number of genes that are core fitness genes}
\item{LeastDependent}{A dataframe where each row corresponds to a gene.There are two columns: \emph{Value} stores the slope of linear model that fits the rank of the gene in allcell lines and the \emph{Gene} stores the gene name}
\item{threshold}{The slope threshold for core fitness genes}
}
\description{
This function identifies the Core Fitness genes from a given Quantative knockout screen dependency matrix where each row is gene and each column the cell line. The function uses all the cell lines and identifies the genes that are essential in majority of the cell lines.
}
\details{
This function implements the idea that if a gene is essential then it should have ranked better in all cell lines including the least dependent cell lines. Instead of calculating rank threshold the ranks are modeled as a linear relation.
For a given gene, we can rank its gene effect score in each cell line, then arrange cell lines in order of increasing gene effect score for that gene. The ranks of the genes in these cell lines are fitted to a linear model where smaller slope values indicates higher dependency in all the cell lines.
A slope threshold is choosen as the minimum density between the two gaussian distributions that are estimated from the distribution of slopes. All genes with a slope less than this threshold is reported. Notice that we do not need to put a constraint such as 90th percentile least depleated cell lines.
}
\examples{
data(exampleSBFData)
results <- ADAM2.SlopeCF(depMat=exampleSBFData)
cfgenes <- results$cfgenes
}
\author{
C. Pacini, E. Karakoc & F. Iorio
}
\keyword{functions}
