% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curated_BAGEL_essential}
\alias{curated_BAGEL_essential}
\title{Predefined essential core set}
\format{A vector of 326 strings
\describe{
  \item{columns}{Gene names}
}}
\usage{
curated_BAGEL_essential
}
\description{
A dataset containing predefined essential genes and it is used for calculating
the true positive rates in ADAM Model
}
\keyword{datasets}
