% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleDepMat}
\alias{exampleDepMat}
\title{Binary dependency map of 17,995 genes for 32 cell lines}
\format{A matrix with 17995 rows and 32 columns:
\describe{
  \item{rows}{Genes}
  \item{columns}{cell lines, samples}
}}
\usage{
exampleDepMat
}
\description{
A dataset containing the binary dependencies where rows are genes and columns are samples.
1 indicates that inactivation of the corresponding gene exerts a significant loss of fitness
in the corresponding cell line, 0 otherwise.
}
\keyword{datasets}
