% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleSBFData}
\alias{exampleSBFData}
\title{Quantitative dependency map of 17,995 genes for 31 cell lines}
\format{A matrix with 17995 rows and 31 columns:
\describe{
  \item{rows}{Genes}
  \item{columns}{cell lines, samples}
}}
\usage{
exampleSBFData
}
\description{
A dataset representing the dependencies where rows are genes and columns are samples.
Each cell is a double value that indicates dependency score where higher values corresponds to
a significant loss of fitness incase of inactivation
}
\keyword{datasets}
